requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langgraph==0.2.76
python-dotenv
faiss-cpu
